pyproject.toml
[project]
name = "engine"
version = "0.1.0"
description = "Video Ad Generator Dependencies"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "anyio>=4.12.1",
    "arq>=0.27.0",
    "asyncpg>=0.31.0",
    "fastapi>=0.134.0",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rich>=14.3.3",
    "sqlalchemy[asyncio]>=2.0.47",
    "tenacity>=9.1.4",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.41.0",
]
